Recommender-Systems/Model/requirements.txt
# uncomment to upload movie images
google-generativeai
streamlit
python-dotenv